+++ verl/trainer/config/_generated_diffusion_trainer.yaml
@@ -42,13 +42,13 @@ actor_rollout_ref:
       forward_only: false
       strategy: fsdp
       dtype: bfloat16
-    _target_: verl.workers.config.FSDPDiffusionActorConfig
+    _target_: verl.workers.config.diffusion.FSDPDiffusionActorConfig
     rollout_n: ${oc.select:actor_rollout_ref.rollout.n,1}
     strategy: fsdp
     ppo_mini_batch_size: 256
     ppo_micro_batch_size_per_gpu: null
     diffusion_loss:
-      _target_: verl.workers.config.DiffusionLossConfig
+      _target_: verl.workers.config.diffusion.DiffusionLossConfig
       loss_mode: flow_grpo
       clip_ratio: 0.0001
       adv_clip_max: 5.0
@@ -89,9 +89,9 @@ actor_rollout_ref:
       forward_only: true
       strategy: fsdp
       dtype: bfloat16
-    _target_: verl.workers.config.FSDPDiffusionActorConfig
+    _target_: verl.workers.config.diffusion.FSDPDiffusionActorConfig
   rollout:
-    _target_: verl.workers.config.DiffusionRolloutConfig
+    _target_: verl.workers.config.diffusion.DiffusionRolloutConfig
     name: ???
     mode: async
     nnodes: 0
@@ -102,6 +102,7 @@ actor_rollout_ref:
     free_cache_engine: true
     tensor_model_parallel_size: 2
     data_parallel_size: 1
+    pipeline_model_parallel_size: 1
     max_num_batched_tokens: 8192
     max_model_len: null
     max_num_seqs: 1024
@@ -113,7 +114,7 @@ actor_rollout_ref:
     engine_kwargs:
       vllm_omni: {}
     val_kwargs:
-      _target_: verl.workers.config.DiffusionSamplingConfig
+      _target_: verl.workers.config.diffusion.DiffusionSamplingConfig
       'n': 1
       num_inference_steps: 40
       seed: 42
@@ -135,7 +136,7 @@ actor_rollout_ref:
     num_inference_steps: 10
     extra_configs: {}
   model:
-    _target_: verl.workers.config.DiffusionModelConfig
+    _target_: verl.workers.config.diffusion.DiffusionModelConfig
     path: ~/models/Qwen/Qwen-Image
     tokenizer_path: null
     use_shm: false
