+++ verl/trainer/config/diffusion/actor/dp_diffusion_actor.yaml
@@ -20,7 +20,7 @@ defaults:
   - _self_
 
 # Target class for this configuration
-_target_: verl.workers.config.FSDPDiffusionActorConfig
+_target_: verl.workers.config.diffusion.FSDPDiffusionActorConfig
 
 # Training strategy: fsdp, fsdp2
 strategy: fsdp
