+++ verl/workers/config/diffusion/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Bytedance Ltd. and/or its affiliates
+# Copyright 2026 Bytedance Ltd. and/or its affiliates
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,25 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import actor, critic, diffusion, engine, model, optimizer, reward, rollout
+from . import actor, model, rollout
 from .actor import *  # noqa: F401
-from .critic import *  # noqa: F401
-from .diffusion import *  # noqa: F401
-from .distillation import *  # noqa: F401
-from .engine import *  # noqa: F401
 from .model import *  # noqa: F401
-from .optimizer import *  # noqa: F401
-from .reward import *  # noqa: F401
 from .rollout import *  # noqa: F401
 
-__all__ = (
-    actor.__all__
-    + critic.__all__
-    + diffusion.__all__
-    + reward.__all__
-    + engine.__all__
-    + optimizer.__all__
-    + rollout.__all__
-    + model.__all__
-    + distillation.__all__
-)
+__all__ = actor.__all__ + model.__all__ + rollout.__all__
