+++ verl/workers/config/model.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
 from dataclasses import dataclass, field
 from typing import Any, Optional
 
@@ -24,7 +23,7 @@ from verl.utils.fs import copy_to_local
 from verl.utils.import_utils import import_external_libs
 from verl.utils.model import get_generation_config, update_model_config
 
-__all__ = ["HFModelConfig", "DiffusionModelConfig", "MtpConfig"]
+__all__ = ["HFModelConfig", "MtpConfig"]
 
 
 @dataclass
@@ -93,7 +92,7 @@ class HFModelConfig(BaseConfig):
     tokenizer_path: Optional[str] = None
     local_tokenizer_path: Optional[str] = None
 
-    # model type, e.g., "language_model", "value_model", "diffusion_model"
+    # model type, e.g., "language_model", "value_model"
     model_type: str = "language_model"
 
     # whether to load tokenizer. This is useful when we only want to load model config
@@ -217,111 +216,6 @@ class HFModelConfig(BaseConfig):
         # per model patch
         if getattr(self.hf_config, "model_type", None) == "kimi_vl":
             self.hf_config.text_config.topk_method = "greedy"
-
-        # Ensure target_modules is a str or list[str] (only if not None)
-        if self.target_modules is not None:
-            if not isinstance(self.target_modules, (str | list)):
-                raise TypeError(
-                    "target_modules must be a string or a list of strings, "
-                    f"but got {type(self.target_modules).__name__}"
-                )
-            if isinstance(self.target_modules, list):
-                for x in self.target_modules:
-                    if not isinstance(x, str):
-                        raise TypeError(
-                            f"All elements in target_modules list must be strings, but found {type(x).__name__}"
-                        )
-
-    def get_processor(self):
-        return self.processor if self.processor is not None else self.tokenizer
-
-
-@dataclass
-class DiffusionModelConfig(BaseConfig):
-    _mutable_fields = {
-        "model_type",
-        "tokenizer_path",
-        "tokenizer",
-        "processor",
-        "local_path",
-        "local_tokenizer_path",
-        "architecture",
-    }
-
-    path: str = MISSING
-    # Handler key matched against @DiffusionModelBase.register(name).
-    # When None (the default), it is auto-detected from model_index.json's ``_class_name``.
-    architecture: Optional[str] = None
-    local_path: Optional[str] = None
-    tokenizer_path: Optional[str] = None
-    local_tokenizer_path: Optional[str] = None
-    model_type: str = "diffusion_model"
-
-    # whether to load tokenizer. This is useful when we only want to load model config
-    load_tokenizer: bool = True
-
-    tokenizer: Any = None
-    processor: Any = None
-
-    # whether to use shared memory
-    use_shm: bool = False
-    trust_remote_code: bool = False
-
-    # custom chat template for the model
-    custom_chat_template: Optional[str] = None
-
-    external_lib: Optional[str] = None
-
-    enable_gradient_checkpointing: bool = True
-
-    # lora related.
-    lora_rank: int = 32
-    lora_alpha: int = 64
-    lora_init_weights: str = "gaussian"
-    target_modules: Optional[Any] = "all-linear"  # allow both "all-linear" and ["q_proj","k_proj"]
-    target_parameters: Optional[list[str]] = None  # for lora adapter on nn.Parameter
-    exclude_modules: Optional[str] = None
-    lora_adapter_path: Optional[str] = None
-
-    # megatron lora config
-    lora: dict[str, Any] = field(default_factory=dict)
-
-    # commonly used sampling parameters for diffusion config.
-    height: int = 512
-    width: int = 512
-    num_inference_steps: int = 10
-
-    # extra configs for algorithm specific features.
-    extra_configs: dict[str, Any] = field(default_factory=dict)
-
-    def __post_init__(self):
-        import_external_libs(self.external_lib)
-        if self.tokenizer_path is None:
-            self.tokenizer_path = os.path.join(self.path, "tokenizer")
-        self.local_path = copy_to_local(self.path, use_shm=self.use_shm)
-
-        if self.architecture is None:
-            import json
-
-            model_index_path = os.path.join(self.local_path, "model_index.json")
-            with open(model_index_path) as f:
-                self.architecture = json.load(f)["_class_name"]
-
-        # construct tokenizer
-        if self.load_tokenizer:
-            self.local_tokenizer_path = copy_to_local(self.tokenizer_path, use_shm=self.use_shm)
-            # Fast tokenizer for diffusion: DiffusionSingleTurnAgentLoop applies chat template on the asyncio
-            # thread (not run_in_executor) so Rust-backed tokenizers avoid RuntimeError: Already borrowed
-            # with recent transformers when combined with thread-pool tokenization.
-            self.tokenizer = hf_tokenizer(
-                self.local_tokenizer_path, trust_remote_code=self.trust_remote_code, use_fast=True
-            )
-            if os.path.exists(os.path.join(self.local_path, "processor")):
-                self.processor = hf_processor(
-                    os.path.join(self.local_path, "processor"), trust_remote_code=self.trust_remote_code
-                )
-            else:
-                self.processor = None
 
         # Ensure target_modules is a str or list[str] (only if not None)
         if self.target_modules is not None:
